post/requirements.txt
matplotlib
pyyaml
docopt